refactor(config): move types in appropriate folders
ref:MANAGER-7865
