+++ packages/manager/modules/config/package.json
@@ -16,9 +16,10 @@
   "sideEffects": false,
   "main": "dist/index.js",
   "module": "dist/index.js",
-  "types": "./dist/types",
+  "types": "./types",
   "files": [
-    "dist"
+    "dist",
+    "types"
   ],
   "scripts": {
     "build": "tsc",
