+++ packages/manager/modules/config/src/environment/environment.constants.ts
@@ -1,4 +1,4 @@
-import { Region } from '../types/ovhLanguages';
+import { Region } from './environment';
 
 export const ALLOWED_REGIONS = [Region.EU, Region.CA, Region.US];
 export const DEFAULT_REGION = Region.EU;
