+++ packages/manager/modules/config/src/locale/index.ts
@@ -1,5 +1,34 @@
-import { LangId, Region } from '../types/ovhLanguages';
 import { LANGUAGES, localeRegex, localeStorageKey } from './locale.constants';
+import { Region } from '../environment/environment';
+
+export const enum CountryCode {
+  US = 'US',
+  CA = 'CA',
+  FR = 'FR',
+  GB = 'GB',
+  ES = 'ES',
+  PL = 'PL',
+  PT = 'PT',
+  IT = 'IT',
+  DE = 'DE',
+}
+export type LangId = 'nl' | 'fr' | 'en' | 'de' | 'es' | 'it' | 'pl' | 'pt';
+
+export interface KeyPairName {
+  name: string;
+  key: string;
+}
+
+export type PreferredRegionByLang = {
+  [langKey in LangId]?: { [regionKey in Region]?: CountryCode };
+};
+
+export interface OVHLanguages {
+  available: Array<KeyPairName>;
+  defaultLoc: string;
+  fallback: string;
+  preferred: PreferredRegionByLang;
+}
 
 const preferredCountry = (language: LangId, region: Region) => {
   if (['FR', 'EN'].includes(language.toUpperCase())) {
