+++ packages/manager/modules/config/src/shims/navigator.d.ts
