+++ packages/manager/modules/config/src/shims/reket.d.ts
