+++ packages/manager/modules/config/tsconfig.json
@@ -12,8 +12,8 @@
     "sourceMap": true,
     "allowJs": true,
     "declaration": true,
-    "declarationDir": "dist/types"
+    "declarationDir": "types"
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "dist", "types"]
 }
